requirements.txt

SpeechRecognition==3.14.3
pyttsx3==2.90
Pyaudio
wikipedia
google-generativeai